Fix CompileJava being realised too soon, and other misc lazy task fixes.
