+++ src/main/java/net/fabricmc/loom/task/RemapJarTask.java
@@ -107,7 +107,7 @@ public abstract class RemapJarTask extends AbstractRemapJarTask {
 		getOptimizeFabricModJson().convention(false).finalizeValueOnRead();
 
 		TaskProvider<NestableJarGenerationTask> processIncludeJars = getProject().getTasks().named(Constants.Task.PROCESS_INCLUDE_JARS, NestableJarGenerationTask.class);
-		getNestedJars().from(getProject().fileTree(processIncludeJars.get().getOutputDirectory()));
+		getNestedJars().from(processIncludeJars.map(task -> getProject().fileTree(task.getOutputDirectory())));
 		getNestedJars().builtBy(processIncludeJars);
 
 		getUseMixinAP().set(LoomGradleExtension.get(getProject()).getMixin().getUseLegacyMixinAp());
